requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
cbor2==5.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
-e git+https://github.com/sunscrapers/djoser.git@610c78c527d4d2c9d2ec834de4df275851200bfd#egg=djoser
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==69.5.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
webauthn==2.1.0
whitenoise==6.6.0
